+++ seanime-desktop/src-tauri/src/tray.rs
@@ -7,11 +7,17 @@ use tauri::{
 
 pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
     let quit_i = MenuItem::with_id(app, "quit", "Quit Seanime", true, None::<&str>)?;
-    // let toggle_visibility_i = MenuItem::with_id(app, "hide", "Hide", true, None::<&str>)?;
+    let restart_i = MenuItem::with_id(app, "restart", "Restart Seanime", true, None::<&str>)?;
+    // let open_web_i = MenuItem::with_id(app, "open_web", "Open Web UI", true, None::<&str>)?;
+    let toggle_visibility_i = MenuItem::with_id(
+        app,
+        "toggle_visibility",
+        "Toggle visibility",
+        true,
+        None::<&str>,
+    )?;
 
-
-    let menu = Menu::with_items(app, &[&quit_i])?;
-
+    let menu = Menu::with_items(app, &[&toggle_visibility_i, &restart_i, &quit_i])?;
 
     let _ = TrayIconBuilder::with_id("tray")
         .icon(app.default_window_icon().unwrap().clone())
@@ -20,6 +26,23 @@ pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
         .on_menu_event(move |app, event| match event.id.as_ref() {
             "quit" => {
                 app.exit(0);
+            }
+            "restart" => app.restart(),
+            "toggle_visibility" => {
+                if let Some(window) = app.get_webview_window(MAIN_WINDOW_LABEL) {
+                    if !window.is_visible().unwrap() {
+                        let _ = window.show();
+                        let _ = window.set_focus();
+                        #[cfg(target_os = "macos")]
+                        app.set_activation_policy(tauri::ActivationPolicy::Regular)
+                            .unwrap();
+                    } else {
+                        let _ = window.hide();
+                        #[cfg(target_os = "macos")]
+                        app.set_activation_policy(tauri::ActivationPolicy::Accessory)
+                            .unwrap();
+                    }
+                }
             }
             // "hide" => {
             //     if let Some(window) = app.get_webview_window(MAIN_WINDOW_LABEL) {
@@ -50,7 +73,8 @@ pub fn create_tray<R: Runtime>(app: &tauri::AppHandle<R>) -> tauri::Result<()> {
                     let _ = window.show();
                     let _ = window.set_focus();
                     #[cfg(target_os = "macos")]
-                    app.set_activation_policy(tauri::ActivationPolicy::Regular).unwrap();
+                    app.set_activation_policy(tauri::ActivationPolicy::Regular)
+                        .unwrap();
                 }
             }
         })